requirements.txt
streamlit
streamlit-webrtc
pillow
google-generativeai
opencv-python
mediapipe